# Conflicts:
#	Gemfile.lock
#	app/models/character.rb
#	app/models/user.rb
#	db/schema.rb

+++ app/controllers/characters_controller.rb
@@ -1,6 +1,11 @@
 class CharactersController < ApplicationController
-  before_action :get_user
-  before_action :get_character, except: [:create]
+  before_action :get_user, except: [:index]
+  before_action :get_character, except: [:index, :create]
+
+  def index
+    @characters = filter_scope
+    render json: @characters, each_serializer: ImageCharacterSerializer
+  end
 
   def show
     set_meta_tags(
@@ -26,6 +31,7 @@ class CharactersController < ApplicationController
 
   def create
     @character = Character.new character_params.merge(user: current_user)
+
     if @character.save
       render json: @character
     else
@@ -34,6 +40,8 @@ class CharactersController < ApplicationController
   end
 
   def update
+    head :unauthorized and return unless @character.managed_by? current_user
+
     if @character.update_attributes character_params
       render json: @character, serializer: CharacterSerializer
     else
@@ -42,6 +50,8 @@ class CharactersController < ApplicationController
   end
 
   def destroy
+    head :unauthorized and return unless @character.managed_by? current_user
+
     if @character.destroy
       render json: @character, serializer: CharacterSerializer
     else
@@ -61,7 +71,8 @@ class CharactersController < ApplicationController
 
   def character_params
     p = params.require(:character).permit(:name, :nickname, :gender, :species, :height, :weight,
-                                          :body_type, :personality, :special_notes, :profile, :likes, :dislikes, :slug)
+                                          :body_type, :personality, :special_notes, :profile, :likes, :dislikes, :slug,
+                                          :shortcode)
 
     if params[:character].include? :color_scheme_attributes
       p[:color_scheme_attributes] = { color_data: params[:character][:color_scheme_attributes][:color_data].permit!, id: @character.color_scheme_id }
@@ -76,5 +87,35 @@ class CharactersController < ApplicationController
     end
 
     p
+  end
+
+  def filter_scope
+    scope = Character.all
+    sort = nil
+    order = nil
+    query = []
+
+    params[:q].split(/\s+/).each do |q|
+      case q
+        when /sort:(\w+)/i
+          sort = $1.downcase
+        when /order:(asc|desc)/i
+          order = $1.upcase
+        else
+          query << q
+      end
+    end
+
+    scope = if %w(created_at updated_at name).include?(sort)
+      scope.order("characters.#{sort} #{order}")
+    else
+      scope.default_order
+    end
+
+    # NSFW / Hidden
+    scope = scope.sfw unless nsfw_on?
+    scope.visible
+
+    scope.search_for(query.join(' '))
   end
 end
